Fixed bye command
