+++ src/main/java/Duke.java
@@ -1,3 +1,4 @@
+
 import java.io.*;
 import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.Files;
@@ -10,19 +11,7 @@ import java.time.LocalDate;
 
 public class Duke {
     public static void main(String[] args) throws IOException {
-        run();
-    }
-
-    public static void run() throws IOException {
-        Ui.startGreeting();
-        Scanner sc = new Scanner(System.in);
-        ArrayList<Task> storeList = new ArrayList<>();
-        Storage.restoreList(storeList);
-        while (sc.hasNextLine()) {
-            String command = sc.nextLine();
-            CommandParser.parseCommand(command, storeList);
-        }
-        sc.close();
+        Ui.run();
     }
 
 }
